Break out custom lint features
