+++ src/editor/codemirror/lint/lint.ts
@@ -34,6 +34,7 @@ import {PanelConstructor, Panel, showPanel, getPanel} from "@codemirror/panel"
 import {gutter, GutterMarker} from "@codemirror/gutter"
 import {RangeSet, Range} from "@codemirror/rangeset"
 import elt from "crelt"
+import { currentlyEditingLine, currentlyEditingLinePlugin } from "./editingLine"
 
 /// Describes a problem or hint for a piece of code.
 export interface Diagnostic {
@@ -729,59 +730,6 @@ function markersForDiagnostics(doc: Text, diagnostics: readonly Diagnostic[]) {
 const lintGutterExtension = gutter({
   class: "cm-gutter-lint",
   markers: view => view.state.field(lintGutterMarkers),
-})
-
-
-
-const currentlyEditingLinePlugin = ViewPlugin.fromClass(class {
-  timeout: any
-
-  update(update: ViewUpdate) {
-    if (!update.docChanged && !update.selectionSet) {
-      return;
-    }
-    console.log("Here2");
-    const mainIndex = update.state.selection?.asSingle()?.ranges[0]?.from;
-    console.log(update.state.selection);
-    if (!mainIndex) {
-      return undefined;
-    }
-    console.log({mainIndex})
-    const doc = update.state.doc;
-    const selectionLine = doc.lineAt(mainIndex).number;
-    update.changes.iterChangedRanges((fromA, toA, fromB, toB) => {
-      if (doc.lineAt(toA).number === selectionLine) {
-        // Woohoo, we changed the line we're on.
-        clearTimeout(this.timeout);
-        setTimeout(() => {
-          update.view.dispatch({ effects: [setEditingLineEffect.of(selectionLine)]});
-        }, 0)
-        this.timeout = setTimeout(() => {
-          update.view.dispatch({ effects: [setEditingLineEffect.of(undefined)]});
-        }, 2000)
-      }
-    })
-  }
-
-  destroy() {
-    clearTimeout(this.timeout)
-  }
-})
-
-export const setEditingLineEffect = StateEffect.define<number|undefined>()
-
-const currentlyEditingLine = StateField.define<number | undefined>({
-  create() {
-    return undefined
-  },
-  update(line, tr) {
-    for (let effect of tr.effects) {
-      if (effect.is(setEditingLineEffect)) {
-        return effect.value;
-      }
-    }
-    return line
-  }
 })
 
 const lintGutterMarkers = StateField.define<RangeSet<GutterMarker>>({
